+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Changed
 - Track mcp.json.
+- mtype resolving now use embedddings.
 
 ### Removed
 - Knowledge graph tools and utils